requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pillow==10.3.0
transformers==4.41.2
torch==2.3.0
easyocr==1.7.1
openai==1.30.1
requests==2.31.0
beautifulsoup4==4.12.3
reportlab==4.1.0
python-dotenv==1.0.1
setuptools
typing-extensions
qrcode[pil]
